third_party/blink/web_tests/virtual/import-maps-without-builtin-modules/external/wpt/import-maps/imported/parsing-specifier-keys.tentative-expected.txt

This is a testharness.js-based test.
PASS Relative URL-like specifier keys / should absolutize strings prefixed with ./, ../, or / into the corresponding URLs
FAIL Relative URL-like specifier keys / should not absolutize strings prefixed with ./, ../, or / with a data: URL base test helper does not support data: base URLs
PASS Relative URL-like specifier keys / should absolutize the literal strings ./, ../, or / with no suffix
PASS Relative URL-like specifier keys / should treat percent-encoded variants of ./, ../, or / as bare specifiers
PASS Absolute URL specifier keys / Accept all absolute URL specifier keys even with fetch schemes as URLs
FAIL Absolute URL specifier keys / should parse absolute URLs, treating unparseable ones as bare specifiers assert_equals: expected "{\"imports\":{\"http://[www.example.com]/\":\"https://base.example/unparseable3/\",\"https://ex ample.org/\":\"https://base.example/unparseable1/\",\"https://example.com/\":\"https://base.example/percentDecoding/\",\"https://example.com/%41\":\"https://base.example/noPercentDecoding\",\"https://example.com///\":\"https://base.example/invalidButParseable2/\",\"https://example.com:demo\":\"https://base.example/unparseable2\",\"https://example.net/\":\"https://base.example/prettyNormal/\",\"https://example.org/\":\"https://base.example/invalidButParseable1/\"},\"scopes\":{}}" but got "{\"imports\":{\"http://[www.example.com]/\":\"https://base.example/unparseable3/\",\"https://ex%20ample.org/\":\"https://base.example/unparseable1/\",\"https://example.com/\":\"https://base.example/percentDecoding/\",\"https://example.com///\":\"https://base.example/invalidButParseable2/\",\"https://example.com/A\":\"https://base.example/noPercentDecoding\",\"https://example.com:demo\":\"https://base.example/unparseable2\",\"https://example.net/\":\"https://base.example/prettyNormal/\",\"https://example.org/\":\"https://base.example/invalidButParseable1/\"},\"scopes\":{}}"
PASS Absolute URL specifier keys / should sort correctly (issue #181)
Harness: the test ran to completion.
